# Conflicts:
#	.ci/scripts/distribution/build-go.sh
#	.ci/scripts/distribution/test-go.sh
#	.ci/scripts/release/compat-update.sh
#	Jenkinsfile
#	bom/pom.xml
#	bpmn-model/pom.xml
#	broker/pom.xml
#	build-tools/pom.xml
#	clients/go/cmd/zbctl/internal/commands/root.go
#	clients/go/go.mod
#	clients/go/go.sum
#	clients/go/internal/containersuite/containerSuite.go
#	clients/go/vendor/github.com/spf13/cobra/.gitignore
#	clients/go/vendor/github.com/spf13/cobra/.travis.yml
#	clients/go/vendor/github.com/spf13/cobra/command.go
#	clients/go/vendor/google.golang.org/grpc/balancer/base/balancer.go
#	clients/go/vendor/google.golang.org/grpc/version.go
#	clients/go/vendor/modules.txt
#	clients/java/pom.xml
#	clients/java/src/main/java/io/zeebe/client/impl/oauth/OAuthCredentialsProvider.java
#	dispatcher/pom.xml
#	dist/pom.xml
#	dist/src/main/config/zeebe.cfg.toml
#	engine/pom.xml
#	exporter-api/pom.xml
#	exporters/elasticsearch-exporter/pom.xml
#	exporters/elasticsearch-exporter/src/main/java/io/zeebe/exporter/ElasticsearchClient.java
#	expression-language/pom.xml
#	gateway-protocol-impl/pom.xml
#	gateway-protocol/pom.xml
#	gateway/pom.xml
#	logstreams/pom.xml
#	msgpack-core/pom.xml
#	msgpack-value/pom.xml
#	parent/pom.xml
#	pom.xml
#	protocol-asserts/pom.xml
#	protocol-impl/pom.xml
#	protocol-test-util/pom.xml
#	protocol/pom.xml
#	qa/integration-tests/pom.xml
#	qa/pom.xml
#	samples/pom.xml
#	test-util/pom.xml
#	test/pom.xml
#	transport/pom.xml
#	upgrade-tests/pom.xml
#	util/pom.xml
#	zb-db/pom.xml

+++ dispatcher/src/test/java/io/zeebe/dispatcher/integration/FragmentBatchIntegrationTest.java
@@ -45,7 +45,7 @@ public final class FragmentBatchIntegrationTest {
   public void init() {
     dispatcher =
         Dispatchers.create("default")
-            .bufferSize(ByteValue.ofKilobytes(32))
+            .bufferSize((int) ByteValue.ofKilobytes(32))
             .actorScheduler(actorSchedulerRule.get())
             .build();
